+++ ambry-router/src/main/java/com/github/ambry/router/TtlUpdateManager.java
@@ -36,9 +36,6 @@ import java.util.List;
 import java.util.Map;
 import java.util.Set;
 import java.util.concurrent.ConcurrentHashMap;
-import java.util.concurrent.ConcurrentMap;
-import java.util.concurrent.atomic.AtomicBoolean;
-import java.util.concurrent.atomic.AtomicLong;
 
 
 /**
@@ -105,14 +102,12 @@ class TtlUpdateManager {
               time.milliseconds(), callback, time, futureResult);
       ttlUpdateOperations.add(ttlUpdateOperation);
     } else {
-      BatchTtlUpdateOperationCallbackTracker tracker =
-          new BatchTtlUpdateOperationCallbackTracker(blobIds, futureResult, callback);
+      BatchOperationCallbackTracker tracker = new BatchOperationCallbackTracker(blobIds, futureResult, callback);
       long operationTimeMs = time.milliseconds();
       for (BlobId blobId : blobIds) {
         TtlUpdateOperation ttlUpdateOperation =
             new TtlUpdateOperation(clusterMap, routerConfig, routerMetrics, blobId, serviceId, expiresAtMs,
-                operationTimeMs, tracker.getCallback(blobId), time,
-                BatchTtlUpdateOperationCallbackTracker.DUMMY_FUTURE);
+                operationTimeMs, tracker.getCallback(blobId), time, BatchOperationCallbackTracker.DUMMY_FUTURE);
         ttlUpdateOperations.add(ttlUpdateOperation);
       }
     }
@@ -219,69 +214,6 @@ class TtlUpdateManager {
         routerMetrics.operationAbortCount.inc();
         routerMetrics.onUpdateBlobTtlError(e);
         NonBlockingRouter.completeOperation(op.getFutureResult(), op.getCallback(), null, e);
-      }
-    }
-  }
-
-  /**
-   * Tracks callbacks for TtlUpdate operations over multiple chunks of a single blob
-   */
-  private static class BatchTtlUpdateOperationCallbackTracker {
-    static final FutureResult<Void> DUMMY_FUTURE = new FutureResult<>();
-
-    private final FutureResult<Void> futureResult;
-    private final Callback<Void> callback;
-    private final long numBlobIds;
-    private final ConcurrentMap<BlobId, Boolean> blobIdToAck = new ConcurrentHashMap<>();
-    private final AtomicLong ackedCount = new AtomicLong(0);
-    private final AtomicBoolean completed = new AtomicBoolean(false);
-
-    /**
-     * Constructor
-     * @param blobIds the {@link BlobId}s being tracked
-     * @param futureResult the {@link FutureResult} to be triggered once acks are received for all blobs
-     * @param callback the {@link Callback} to be triggered once acks are received for all blobs
-     */
-    BatchTtlUpdateOperationCallbackTracker(List<BlobId> blobIds, FutureResult<Void> futureResult,
-        Callback<Void> callback) {
-      numBlobIds = blobIds.size();
-      blobIds.forEach(blobId -> blobIdToAck.put(blobId, false));
-      if (blobIdToAck.size() != numBlobIds) {
-        throw new IllegalArgumentException("The list of BlobIds provided has duplicates: " + blobIds);
-      }
-      this.futureResult = futureResult;
-      this.callback = callback;
-    }
-
-    /**
-     * Gets a {@link Callback} personalized for {@code blobId}.
-     * @param blobId the {@link BlobId} for which the
-     * @return the {@link Callback} to be used with the {@link TtlUpdateOperation} for {@code blobId}.
-     */
-    Callback<Void> getCallback(final BlobId blobId) {
-      return (result, exception) -> {
-        if (exception == null) {
-          if (blobIdToAck.put(blobId, true)) {
-            // already acked once
-            complete(new RouterException("Ack for " + blobId + " arrived more than once",
-                RouterErrorCode.UnexpectedInternalError));
-          } else if (ackedCount.incrementAndGet() >= numBlobIds) {
-            // acked for the first time for this blob id and all the blob ids have been acked
-            complete(null);
-          }
-        } else {
-          complete(exception);
-        }
-      };
-    }
-
-    /**
-     * Completes the batch operation
-     * @param e the {@link Exception} that occurred (if any).
-     */
-    private void complete(Exception e) {
-      if (completed.compareAndSet(false, true)) {
-        NonBlockingRouter.completeOperation(futureResult, callback, null, e, false);
       }
     }
   }
